Item Database. Inventory UI, Drag and Drop function
I incorporated a database of item objects to act as an inventory. Using this I made an adaptable UI element that displays the elements within the inventory. Furthermore I incorporated a drag and drop element where the items can be dragged and dropped into predetermined containers. The crafting system will work by having the user drag items into specified containers that determine which attribute it'll use when crafting.