Crafting/Inventory Updating
The crafting system now gives the player the ability to move objects into the crafting system and attempt to craft an object. If successful the inventory will delete the used item and add the crafted one. THERE are bugs where if the user removes the item after they already place them onto the crafting table. A bug that is all on how the data is transferred between the scripts and can be fixed. I'll get to work on the fixes on my next push.